importScrollView from react-native-gesture-handler
IMPORTANT!!! import ScrollView from react-native-gesture-handler instead just react-native. this mode allows the user - even if the TextTicker is already inside a ScrollView (horizontal) - to grab the text and drag (with the 'scroll' property active). Without the gesture-handler, this is not possible, the drag does not work (only scroll the FATHER ScrollView, the son doesn't scroll).
Video showing the problem and after fixed:
https://www.youtube.com/watch?v=Ev1_tT98kDI&ab_channel=ManoeldeOliveiraPradoNeto
Project as an example in the following repository:
https://github.com/ManoelPradoMark22/ignite_reactNative_gofinances_project2
My commit when I add the react-native-text-ticker lib:
ManoelPradoMark22/ignite_reactNative_gofinances_project2@0f62903